requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9
pyyaml>=6.0
httpx>=0.25.0
upstash-redis>=1.0.0
qstash>=3.0.0
upstash-vector>=0.1.0
# upstash-workflow requires qstash<3; omit and use QStash HTTP only for replay
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-xai>=0.1.0
daytona-sdk>=0.1.0
